% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_over & Mutation.R
\name{crossover}
\alias{crossover}
\title{Cross over}
\usage{
crossover(parents, dat, fitfunc = "AIC", family = "gaussian")
}
\arguments{
\item{parents}{list, parents to generate offsprings,length=2}

\item{dat}{dataframe, data to generate offspring(default sets the dependent variable in first column and independent varialbes is other columns)}

\item{fitfunc}{method, model selection method(default is AIC)}

\item{family}{family,for linear regression model the family should be a continuous probability density function (default is gaussian family)}
}
\description{
Generate offsprings by randomly pick the position of chromosomes
}
