% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper Function.R
\name{getFitness}
\alias{getFitness}
\title{getFitness}
\usage{
getFitness(genomes)
}
\arguments{
\item{genomes}{list, composed of chromosomes and fitness value with lenght equal to population size(default=NULL)}
}
\value{
vector, composed of AIC of each chromsome
}
\description{
Retrieves AICs from list of chromosomes
}
