% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_over & Mutation.R
\name{mutateChrom}
\alias{mutateChrom}
\title{Muatation}
\usage{
mutateChrom(chrom, nMutate, dat, fitfunc = "AIC", family = "gaussian")
}
\arguments{
\item{chrom}{vector, a chromsome which a mutation is desired}

\item{nMutate}{integer, indicating how many mutations to perform on chromosome}

\item{dat}{dataframe, data to do mutation(default sets the dependent variable in first column and independent varialbes is other columns)}

\item{fitfunc}{method, model selection method(default is AIC)}

\item{family}{family,for linear regression model the family should be a continuous probability density function (default is gaussian family)}
}
\description{
Performs mutation on single chromosomes
}
