% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross_over & Mutation.R
\name{mutatePop}
\alias{mutatePop}
\title{Muatation}
\usage{
mutatePop(pop, nMutations, fitfunc = "AIC", family = "gaussian")
}
\arguments{
\item{pop}{list, a population which mutation is desired}

\item{fitfunc}{method, model selection method(default is AIC)}

\item{family}{family,for linear regression model the family should be a continuous probability density function (default is gaussian family)}

\item{nMutate}{integer, indicating how many mutations to perform on chromosome}

\item{dat}{dataframe, data to do mutation(default sets the dependent variable in first column and independent varialbes is other columns) )}
}
\description{
Perform mutations on population
}
