% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Next Generation.R
\name{nextGen}
\alias{nextGen}
\title{nextGen}
\usage{
nextGen(pop, pSelect = 0.2, pMutate = 0.01, fitfunc = "AIC",
  family = "gaussian", end_iter_cond = F)
}
\arguments{
\item{pop}{list, population}

\item{pSelect}{float, the worse part to remove from the population(default=0.2)}

\item{pMutate}{float, the probability used to decide number of mutation(default=0.01)}

\item{fitfunc}{method, model selection method(default is AIC)}

\item{family}{family,for linear regression model the family should be a continuous probability density function (default is gaussian family)}

\item{end_iter_cond}{logical, the index to decide whether to stop iteration (default=F)}
}
\value{
list, population
}
\description{
Make the next generation
}
